2019 Day 9.

+++ src/main/java/eu/janvdb/aoc2019/day2/Day2.java
@@ -4,7 +4,7 @@ import eu.janvdb.aoc2019.common.Computer;
 
 public class Day2 {
 
-	private static final int[] INPUT = {
+	private static final long[] INPUT = {
 			1, 0, 0, 3, 1, 1, 2, 3, 1, 3, 4, 3, 1, 5, 0, 3, 2, 13, 1, 19, 1, 9, 19, 23, 2, 23, 13, 27, 1, 27, 9, 31, 2,
 			31, 6, 35, 1, 5, 35, 39, 1, 10, 39, 43, 2, 43, 6, 47, 1, 10, 47, 51, 2, 6, 51, 55, 1, 5, 55, 59, 1, 59, 9,
 			63, 1, 13, 63, 67, 2, 6, 67, 71, 1, 5, 71, 75, 2, 6, 75, 79, 2, 79, 6, 83, 1, 13, 83, 87, 1, 9, 87, 91, 1,
@@ -33,7 +33,7 @@ public class Day2 {
 		while (true) {
 			int noun = 0;
 			while (noun <= verb) {
-				int result = new Computer(INPUT).runWithInput(verb, noun);
+				long result = new Computer(INPUT).runWithInput(verb, noun);
 				if (result == EXPECTED_RESULT) {
 					System.out.printf("%d/%d -> %d", verb, noun, 100 * verb + noun);
 					return;
